Rewrite Groovy exception code as Java
Groovy has a strange bug.  When you compile a library with JDK6 byte
code and try to run it on a JDK7 or higher JVM; the resulting error is
mismatched arguments for exceptions.

The problem is fully described in:
http://blog.proxerd.pl/article/how-to-fix-incompatibleclasschangeerror-for-your-groovy-projects-running-on-jdk7

The solution to avoid future incompatibilities is to rewrite the Groovy
exceptions as Java exceptions.

fixes #82