added getFullBranchRegexString to support pipelines
Pipeline branch filters are additive.  This means if I specify both a
wildcard and a regex filter, then it will only match if a branch matches
both filters.

For this reason, when using regex filter, we need to build a regex
combining branch names with regex.  The combined filter will be a regex
including all so that a branch may match either (name or regex).