StaticMocking should always create request files
even if empty.  While testing in practice I was using a HEAD method first
to check it didn't exist only to later use a POST method on the same
endpoint to create it.  This caused issues when testing with the mocks
because the file didn't exist i.e. mock 404 not found.