Force 1 hash iteration for password AES cipher
Due to the way the initilization vector (IV) is created in my code with a
password checksum, I am requiring that it pass through PBKDF2 with HMAC
SHA-256 with variable iterations function.  This means the minimum
iteration allowed is 1 because at 0 no key derivation occurs.

If the IV were restricted to SHA-256 characters (base16) then the number
of possibilities for the IV would be 16^16.  This is drastically less
than 2^128.  For this reason, the additional PBKDF layer for the IV is
necessary.